+++ CHANGELOG.md
@@ -8,11 +8,13 @@
   - com.dwarfeng.statistics.impl.handler.receiver.DoNothingReceiver。
   - com.dwarfeng.statistics.impl.handler.receiver.InjvmReceiver。
   - com.dwarfeng.statistics.impl.handler.receiver.KafkaReceiver。
+  - com.dwarfeng.statistics.impl.handler.receiver.DubboReceiver。
 
 - 实现预设调度器。
   - com.dwarfeng.statistics.impl.handler.dispatcher.DrainDispatcher。
   - com.dwarfeng.statistics.impl.handler.dispatcher.InjvmDispatcher。
   - com.dwarfeng.statistics.impl.handler.dispatcher.KafkaDispatcher。
+  - com.dwarfeng.statistics.impl.handler.dispatcher.DubboDispatcher。
 
 - 实现预设驱动器。
   - com.dwarfeng.statistics.impl.handler.driver.CronDriverProvider。